# performance patch of JsonConverterHelper.cs