@AlexGhiondea 2022-01-19
@PiDiBi thanks for your PR!

Do you happen to have the benchmarks for allocations? Did they change?