@PiDiBi 2022-03-18
should I delete this PR? nobody interested in performance improvements?